Detaching augmentation from the entrypoint to enable more entities.
